move intent tests to file
